Rename the Video class to GstVideo, and add in a new Video typedef for it in a separate header. This allows for two scenarios: Clients that want multiple video playback systems simultaneously, and clients that want a single system they can easily replace.

+++ projects/video/gstreamer-1.0/src/gstreamer/video_meta_cache.cpp
@@ -183,7 +183,7 @@ bool VideoMetaCache::getVideoInfo(const std::string& path, float& outDuration, i
 	} else {
 		int hours = atoi(duration.substr(0,1).c_str());
 		int minutes = atoi(duration.substr(2,2).c_str());
-		float seconds = atof(duration.substr(5).c_str());
+		float seconds = static_cast<float>(atof(duration.substr(5).c_str()));
 		outDuration = static_cast<float>(hours * 60 * 60) + static_cast<float>(minutes * 60) + seconds;
 	}
